:recycle: split json module into package
