+++ qalib/translators/json/embed.py
@@ -1,0 +1,83 @@
+from datetime import datetime
+from typing import Optional
+
+import discord.types.embed
+
+from qalib.translators.element.embed import EmbedAdapter
+from qalib.translators.element.expansive import ExpansiveEmbedAdapter
+from qalib.translators.element.types.embed import EmbedBaseAdapter, Colour, make_colour
+from qalib.translators.json import components
+
+
+class JSONEmbedBaseAdapter(EmbedBaseAdapter):
+
+    def __init__(self, embed: components.Embed):
+        self._embed = embed
+
+    @property
+    def title(self) -> str:
+        return self._embed["title"]
+
+    @property
+    def description(self) -> Optional[str]:
+        return self._embed.get("description")
+
+    @property
+    def type(self) -> discord.types.embed.EmbedType:
+        return self._embed["type"] if "type" in self._embed else "rich"
+
+    @property
+    def colour(self) -> Colour | int:
+        return make_colour(self._embed["colour"] if "colour" in self._embed else self._embed["color"])
+
+    @property
+    def timestamp(self) -> Optional[datetime]:
+        """Method to render a timestamp element into a datetime object
+
+        Returns (Optional[datetime]): A datetime object
+        """
+        timestamp = self._embed.get("timestamp")
+        if timestamp is None:
+            return None
+
+        date = timestamp["date"]
+        date_format = timestamp.get("format", "%Y-%m-%d %H:%M:%S.%f")
+        return datetime.strptime(date, date_format)
+
+    @property
+    def footer(self) -> Optional[components.Footer]:
+        return self._embed.get("footer")
+
+    @property
+    def image(self) -> Optional[str]:
+        return self._embed.get("image")
+
+    @property
+    def thumbnail(self) -> Optional[str]:
+        return self._embed.get("thumbnail")
+
+    @property
+    def author(self) -> Optional[components.Author]:
+        return self._embed.get("author")
+
+
+class JSONEmbedAdapter(JSONEmbedBaseAdapter, EmbedAdapter):
+
+    @property
+    def fields(self) -> list[components.Field]:
+        return self._embed.get("fields", [])
+
+
+class JSONExpansiveEmbedAdapter(JSONEmbedBaseAdapter, ExpansiveEmbedAdapter):
+    _embed: components.ExpansiveEmbed
+
+    def __init__(self, embed: components.ExpansiveEmbed):
+        super().__init__(embed)
+
+    @property
+    def field(self) -> components.Field:
+        return self._embed["field"]
+
+    @property
+    def page_number_key(self) -> Optional[str]:
+        return self._embed.get("page_number_key")
